// ui/lib/api-client/models/RobotsAnalysisRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Agent Navigability Simulator API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RobotsAnalysisRequest
 */
export interface RobotsAnalysisRequest {
    /**
     * 
     * @type {string}
     * @memberof RobotsAnalysisRequest
     */
    url: string;
}

/**
 * Check if a given object implements the RobotsAnalysisRequest interface.
 */
export function instanceOfRobotsAnalysisRequest(value: object): value is RobotsAnalysisRequest {
    if (!('url' in value) || value['url'] === undefined) return false;
    return true;
}

export function RobotsAnalysisRequestFromJSON(json: any): RobotsAnalysisRequest {
    return RobotsAnalysisRequestFromJSONTyped(json, false);
}

export function RobotsAnalysisRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): RobotsAnalysisRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'url': json['url'],
    };
}

export function RobotsAnalysisRequestToJSON(json: any): RobotsAnalysisRequest {
    return RobotsAnalysisRequestToJSONTyped(json, false);
}

export function RobotsAnalysisRequestToJSONTyped(value?: RobotsAnalysisRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'url': value['url'],
    };
}