// ui/lib/api-client/models/RunRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Agent Navigability Simulator API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RunRequest
 */
export interface RunRequest {
    /**
     * 
     * @type {string}
     * @memberof RunRequest
     */
    taskType: string;
    /**
     * 
     * @type {string}
     * @memberof RunRequest
     */
    prompt?: string | null;
    /**
     * 
     * @type {string}
     * @memberof RunRequest
     */
    url: string;
    /**
     * 
     * @type {string}
     * @memberof RunRequest
     */
    model: string;
    /**
     * 
     * @type {string}
     * @memberof RunRequest
     */
    env?: string;
    /**
     * 
     * @type {string}
     * @memberof RunRequest
     */
    stealth?: RunRequestStealthEnum;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof RunRequest
     */
    proxies?: { [key: string]: any; } | null;
}


/**
 * @export
 */
export const RunRequestStealthEnum = {
    None: 'none',
    Basic: 'basic',
    Aggressive: 'aggressive'
} as const;
export type RunRequestStealthEnum = typeof RunRequestStealthEnum[keyof typeof RunRequestStealthEnum];


/**
 * Check if a given object implements the RunRequest interface.
 */
export function instanceOfRunRequest(value: object): value is RunRequest {
    if (!('taskType' in value) || value['taskType'] === undefined) return false;
    if (!('url' in value) || value['url'] === undefined) return false;
    if (!('model' in value) || value['model'] === undefined) return false;
    return true;
}

export function RunRequestFromJSON(json: any): RunRequest {
    return RunRequestFromJSONTyped(json, false);
}

export function RunRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): RunRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'taskType': json['task_type'],
        'prompt': json['prompt'] == null ? undefined : json['prompt'],
        'url': json['url'],
        'model': json['model'],
        'env': json['env'] == null ? undefined : json['env'],
        'stealth': json['stealth'] == null ? undefined : json['stealth'],
        'proxies': json['proxies'] == null ? undefined : json['proxies'],
    };
}

export function RunRequestToJSON(json: any): RunRequest {
    return RunRequestToJSONTyped(json, false);
}

export function RunRequestToJSONTyped(value?: RunRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'task_type': value['taskType'],
        'prompt': value['prompt'],
        'url': value['url'],
        'model': value['model'],
        'env': value['env'],
        'stealth': value['stealth'],
        'proxies': value['proxies'],
    };
}