// ui/lib/api-client/models/RunResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Agent Navigability Simulator API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RunResponse
 */
export interface RunResponse {
    /**
     * 
     * @type {string}
     * @memberof RunResponse
     */
    runId: string;
    /**
     * 
     * @type {string}
     * @memberof RunResponse
     */
    sessionLiveViewUrl: string;
    /**
     * 
     * @type {string}
     * @memberof RunResponse
     */
    sessionId: string;
}

/**
 * Check if a given object implements the RunResponse interface.
 */
export function instanceOfRunResponse(value: object): value is RunResponse {
    if (!('runId' in value) || value['runId'] === undefined) return false;
    if (!('sessionLiveViewUrl' in value) || value['sessionLiveViewUrl'] === undefined) return false;
    if (!('sessionId' in value) || value['sessionId'] === undefined) return false;
    return true;
}

export function RunResponseFromJSON(json: any): RunResponse {
    return RunResponseFromJSONTyped(json, false);
}

export function RunResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): RunResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'runId': json['run_id'],
        'sessionLiveViewUrl': json['session_live_view_url'],
        'sessionId': json['session_id'],
    };
}

export function RunResponseToJSON(json: any): RunResponse {
    return RunResponseToJSONTyped(json, false);
}

export function RunResponseToJSONTyped(value?: RunResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'run_id': value['runId'],
        'session_live_view_url': value['sessionLiveViewUrl'],
        'session_id': value['sessionId'],
    };
}